Ajout de question de cpp rapide modification sur la question de l'optimisation du calcul de distance Enlevé le projet non nécessaire Déplacé le code minimaliste de Q4 et Q5 dans le test Updaté Q5 en ajoutant les question J'ai remis le projet Q4 et Q5 dans le framework de vs17 au lieu de vs19. Il y avait un mix de framework entre les projets
