--- Cpp/Q4/Cpp.Q4.Core/VectorFilter.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include <memory>
-
