+++ Cpp/Q4/Cpp.Q4.Tests/Cpp.Q4.Tests.vcxproj
@@ -15,7 +15,7 @@
     <ProjectGuid>{CE517E47-AAFB-4B88-B10C-44533492F8B9}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
     <RootNamespace>CppQ4Tests</RootNamespace>
-    <WindowsTargetPlatformVersion>10.0.18362.0</WindowsTargetPlatformVersion>
+    <WindowsTargetPlatformVersion>10.0.17763.0</WindowsTargetPlatformVersion>
     <ProjectSubType>NativeUnitTestProject</ProjectSubType>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -97,7 +97,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
-    <ClInclude Include="VectorFilter.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
