+++ Cpp/Q4/Cpp.Q4.Tests/Cpp.Q4.Tests.vcxproj.filters
@@ -26,8 +26,5 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="VectorFilter.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
 </Project>
