+++ Cpp/Q4/Cpp.Q4.Tests/VectorFilterTest.cpp
@@ -1,12 +1,24 @@
 #include "pch.h"
 #include "CppUnitTest.h"
-
-#include "VectorFilter.h"
+#include <exception>
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
 namespace CppQ4Tests
 {
+
+    class VectorFilter
+    {
+    public:
+        static bool is_within_range(const int value, int min_limit_inclusive, int max_limit_exclusive)
+        {
+            if (min_limit_inclusive >= max_limit_exclusive)
+                throw std::exception("Invalid min/max combination.");
+
+            return (value >= min_limit_inclusive && value < max_limit_exclusive);
+        }
+    };
+
 	TEST_CLASS(VectorFilterTest)
 	{
 	public:
