--- Cpp/Q5/Cpp.Q5.Tests/AdvancedMath.h
@@ -1,17 +1,0 @@
-#pragma once
-
-class AdvancedMath
-{
-    int mean(int* array, int Size)
-    {
-        int i = 0;
-        int Mean = 0;
-        while (1)
-        {
-            Mean = Mean + array[i];
-            if(i >= Size)
-                break;
-        }
-        return Mean / i;
-    }
-};
