+++ Cpp/Q5/Cpp.Q5.Tests/Cpp.Q5.Tests.vcxproj
@@ -15,7 +15,7 @@
     <ProjectGuid>{BCE77134-005D-41CD-A9DA-CA45C84E1C57}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
     <RootNamespace>CppQ5Tests</RootNamespace>
-    <WindowsTargetPlatformVersion>10.0.18362.0</WindowsTargetPlatformVersion>
+    <WindowsTargetPlatformVersion>10.0.17763.0</WindowsTargetPlatformVersion>
     <ProjectSubType>NativeUnitTestProject</ProjectSubType>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -96,7 +96,6 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="AdvancedMath.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
