+++ Cpp/Q5/Cpp.Q5.Tests/Cpp.Q5.Tests.vcxproj.filters
@@ -26,8 +26,5 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="AdvancedMath.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
 </Project>
