+++ InterviewQuestions.sln
@@ -19,8 +19,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Cpp.Q3.Tests", "Cpp\Q3\Cpp.
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Q4", "Q4", "{FAE8702B-2479-46F5-B576-FBE7A49D884F}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Cpp.Q4.Core", "Cpp\Q4\Cpp.Q4.Core\Cpp.Q4.Core.vcxproj", "{8CD46EF0-1822-4D29-8184-2B07AD1F9245}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Cpp.Q4.Tests", "Cpp\Q4\Cpp.Q4.Tests\Cpp.Q4.Tests.vcxproj", "{CE517E47-AAFB-4B88-B10C-44533492F8B9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Q5", "Q5", "{95BE2B0D-E967-4FE5-8F38-9959608E8201}"
@@ -45,10 +43,6 @@ Global
 		{8B094553-964E-4458-935A-AAF9C4E15D2C}.Debug|x64.Build.0 = Debug|x64
 		{8B094553-964E-4458-935A-AAF9C4E15D2C}.Release|x64.ActiveCfg = Release|x64
 		{8B094553-964E-4458-935A-AAF9C4E15D2C}.Release|x64.Build.0 = Release|x64
-		{8CD46EF0-1822-4D29-8184-2B07AD1F9245}.Debug|x64.ActiveCfg = Debug|x64
-		{8CD46EF0-1822-4D29-8184-2B07AD1F9245}.Debug|x64.Build.0 = Debug|x64
-		{8CD46EF0-1822-4D29-8184-2B07AD1F9245}.Release|x64.ActiveCfg = Release|x64
-		{8CD46EF0-1822-4D29-8184-2B07AD1F9245}.Release|x64.Build.0 = Release|x64
 		{CE517E47-AAFB-4B88-B10C-44533492F8B9}.Debug|x64.ActiveCfg = Debug|x64
 		{CE517E47-AAFB-4B88-B10C-44533492F8B9}.Debug|x64.Build.0 = Debug|x64
 		{CE517E47-AAFB-4B88-B10C-44533492F8B9}.Release|x64.ActiveCfg = Release|x64
@@ -69,7 +63,6 @@ Global
 		{949FE219-E5B8-473D-8BA3-87E4813DC744} = {56E0B380-7657-4318-A22C-6AB51BD40D4C}
 		{8B094553-964E-4458-935A-AAF9C4E15D2C} = {949FE219-E5B8-473D-8BA3-87E4813DC744}
 		{FAE8702B-2479-46F5-B576-FBE7A49D884F} = {56E0B380-7657-4318-A22C-6AB51BD40D4C}
-		{8CD46EF0-1822-4D29-8184-2B07AD1F9245} = {FAE8702B-2479-46F5-B576-FBE7A49D884F}
 		{CE517E47-AAFB-4B88-B10C-44533492F8B9} = {FAE8702B-2479-46F5-B576-FBE7A49D884F}
 		{95BE2B0D-E967-4FE5-8F38-9959608E8201} = {56E0B380-7657-4318-A22C-6AB51BD40D4C}
 		{BCE77134-005D-41CD-A9DA-CA45C84E1C57} = {95BE2B0D-E967-4FE5-8F38-9959608E8201}
